requirements.txt
Flask
Flask-Bcrypt
Flask-Login
Flask-Markdown
Flask-Script
Flask-SQLAlchemy
Flask-WTF
Unidecode
